Remove semantic and datatables
When scaling the AAF DS to 2000+ international IdP the result was less
than performant. The browser window would effectively lockup for 20+
seconds on page load.

On working extensively with Chrome developer tools it is quite evident
that semantic-ui and datatables in unison are a major component of the
problem.

This commit removes the overall javascript and css bloat, additional
commits will refine html and integrate much more lightweight solutions
to the challenges DS has to solve.