Ensure DS experience on mobile is useful
Previously the continue button would be rendered well off the page
making the DS useless on smaller mobile phones.

This change sets height to be a proportion of viewport units, not pixels
which is well supported. https://caniuse.com/#feat=viewport-units. This
has been further confirmed with browserstack.

As part of this minor tweaks to widths/heights of other components have
been made, often to fit text on a single line at smaller resolutions.