Unify SP response generation path
Ensure that all redirections to service providers are handled by a
single code path so that we can later introduce verifications for the
return path per the sstc-saml-idp-discovery specification.

It is possible that no IdP has been selected if passive has been
requested hence the additional boolean check that is also introduced
here.