Rename function on entity_cache
Previous naming was not exactly clear as to intended usage. New naming
is closer to what the SAML spec intends, that is a default discovery
service determined by either a default flag or ordering of entry.

This will be complimented by a new function that provides access to all
defined discovery response endpoints associated with an SPSSODescriptor.