Implement return url validation
This implementation ensures that this DS impl abides by the
Identity Provider Discovery Service Protocol and Profile, section 2.5.
That is ensuring that the DS will not send a user to an unregistered
return URL.

This logic will need to be revisited at a future date to remove the
ability to disable it. It exists in this form now so that existing DS
deployments can act in a permissive mode and acquire logging about any
potential impact enabling this may have.

Several RuboCop metrics have been disabled for this function, some may
be removed by the refactoring required above at a later date, the
remained I don't consider offer any value in this specific case, PR may
of course disagree at which point i'll revisit.

Closes #100.