Add error page for 403 response
When users of services with invalid `return_url` triggered the new
verification code they would be presented with a white page which was
not good UX by any measure.

They will now be provided a much more specific message, with the common
cause clearly identified so the metadata can be adjusted.