Provide error screen for unknown service
Previously users hitting the DS from a service that is unknown to the
federation based on entityID would receive an error indicating that the
return parameter was invalid.

AAF support had indicated this was sub-optimal especially for new
registrations.

This change provides a specific error message in the case the
destination service is unknown and provides advice for the service
registrants to retry the discovery process after sufficient time has
passed post approval.