Prevent failure when UTF8 provided in UserAgent
When receiving UserAgent values with UTF8 content the following error
was occured, resulting in end users acquiring a 500 response:

```
2020-08-28 08:03:00 - Encoding::UndefinedConversionError - "\xC3" from ASCII-8BIT to UTF-8:
        .rbenv/versions/2.7.1/lib/ruby/2.7.0/json/common.rb:224:in `generate'
        .rbenv/versions/2.7.1/lib/ruby/2.7.0/json/common.rb:224:in `generate'
        discovery-service/lib/discovery_service/auditing.rb:33:in `record_entry'
```

UTF8 in http headers seems to be an edge case at best and possibly
breaks RFC, but regardless we should be accepting it without error.

Result here is to strip any UTF8 we are provided with, this is fine for
our purposes as we only store UA to assist in debugging.